Carve out resolveExpectedResponseProperties() function.
