+++ lib/unexpectedExpress.js
@@ -5,27 +5,8 @@ const _ = require('underscore');
 const FormData = require('form-data');
 const qs = require('qs');
 const messy = require('messy');
-const metadataPropertyNames = [
-  'strictAsync',
-  'errorPassedToNext',
-  'isDestroyed',
-  'nextCalled',
-  'locals',
-  'url',
-  'requestDestroyed'
-];
 
-const responsePropertyNames = messy.HttpResponse.propertyNames.concat(
-  metadataPropertyNames
-);
-
-function hasKeys(x) {
-  return Object.keys(x).length > 0;
-}
-
-function validateResponseProperties(x) {
-  return _.intersection(Object.keys(x), responsePropertyNames).length > 0;
-}
+const resolveExpectedResponseProperties = require('./resolveExpectedResponseProperties');
 
 module.exports = {
   name: 'unexpected-express',
@@ -78,17 +59,14 @@ module.exports = {
         }
 
         const context = {};
-
         const nextCalls = [];
 
-        value = _.extend({}, value);
         const requestProperties =
           typeof value.request === 'string'
             ? { url: value.request }
             : _.extend({}, value.request);
 
         let requestBody = requestProperties.body;
-
         let requestForm = requestProperties.form;
 
         const httpRequest = new messy.HttpRequest({
@@ -132,6 +110,7 @@ module.exports = {
           requestBody = httpRequest.body;
         }
 
+        value = { ...value };
         delete value.request;
         delete requestProperties.method;
         delete requestProperties.url;
@@ -198,53 +177,10 @@ module.exports = {
           delete requestProperties.query;
         }
 
-        const responseProperties = value.response;
-        delete value.response;
-        let expectedResponseProperties;
-
-        if (typeof responseProperties === 'number') {
-          expectedResponseProperties = { statusCode: responseProperties };
-        } else if (
-          typeof responseProperties === 'string' ||
-          Buffer.isBuffer(responseProperties)
-        ) {
-          expectedResponseProperties = { body: responseProperties };
-        } else if (Array.isArray(responseProperties)) {
-          throw new Error(
-            'unexpected-express: Response object must be a number, string, buffer or object.'
-          );
-        } else {
-          if (
-            responseProperties &&
-            hasKeys(responseProperties) &&
-            !validateResponseProperties(responseProperties)
-          ) {
-            throw new Error(
-              'unexpected-express: Response object specification incomplete.'
-            );
-          }
-
-          expectedResponseProperties = _.extend({}, responseProperties);
-        }
-
-        const expectedMetadata = _.extend(
-          {},
-          _.pick(expectedResponseProperties, metadataPropertyNames),
-          _.pick(value, metadataPropertyNames)
-        );
-        expectedResponseProperties = _.omit(
+        const {
           expectedResponseProperties,
-          metadataPropertyNames
-        );
-
-        const missingResponseProperties = Object.keys(
-          expectedResponseProperties
-        ).filter(key => responsePropertyNames.indexOf(key) === -1);
-        if (missingResponseProperties.length > 0) {
-          throw new Error(
-            `Property "${missingResponseProperties[0]}" does not exist on the response object.`
-          );
-        }
+          expectedMetadata
+        } = resolveExpectedResponseProperties(value);
 
         const req = new http.IncomingMessage({
           destroy() {
@@ -523,6 +459,7 @@ module.exports = {
                 )
               );
             });
+
             return expect.promise.settle(promiseByKey).then(promises => {
               if (promises.some(promise => promise.isRejected())) {
                 expect.fail({
